refactor - class rename
 EquationFilter -> EquationFactory
 StringSplitter -> RawEquationSplitter
