+++ src/main/java/calculator/spliter/RawEquationSplitter.java
@@ -3,7 +3,7 @@ package calculator.spliter;
 import java.util.Arrays;
 import java.util.List;
 
-public class StringSplitter {
+public class RawEquationSplitter {
     private static final String BLANK = " ";
     private static final String NUMBER_FORMAT = "\\d+";
 
@@ -15,12 +15,16 @@ public class StringSplitter {
 
     private static void validate(String[] equation) {
         for (int i = 0; i < equation.length; i++) {
-            if (i % 2 == 0 && isNotNumber(equation[i])) {
-                throw new IllegalArgumentException("연산자의 위치가 잘못되었습니다.");
-            }
-            if (i % 2 != 0 && isNumber(equation[i])) {
-                throw new IllegalArgumentException("숫자의 위치가 잘못되었습니다.");
-            }
+            checkPosition(equation, i);
+        }
+    }
+
+    private static void checkPosition(String[] equation, int i) {
+        if (i % 2 != 0 && isNumber(equation[i])) {
+            throw new IllegalArgumentException("숫자의 위치가 잘못되었습니다.");
+        }
+        if (i % 2 == 0 && isNotNumber(equation[i])) {
+            throw new IllegalArgumentException("연산자의 위치가 잘못되었습니다.");
         }
     }
